redirection to detail coins page from trending home coins done

+++ app/src/main/java/com/crypto/prices/model/crypto/search/Ath.kt
@@ -47,7 +47,7 @@ data class Ath(
     val pln: BigDecimal,
     val rub: BigDecimal,
     val sar: BigDecimal,
-    val sats: Int,
+    val sats: BigDecimal,
     val sek: BigDecimal,
     val sgd: BigDecimal,
     val thb: BigDecimal,
